added interfaces for services

+++ src/main/java/com/vriddhi/tablebookingapp/service/UserService.java
@@ -1,59 +1,63 @@
 package com.vriddhi.tablebookingapp.service;
 
 import com.vriddhi.tablebookingapp.dto.UserResponseDTO;
+import com.vriddhi.tablebookingapp.exception.ResourceNotFoundException;
 import com.vriddhi.tablebookingapp.model.User;
 import com.vriddhi.tablebookingapp.repository.UserRepository;
+import io.swagger.v3.oas.annotations.Operation;
 import lombok.extern.slf4j.Slf4j;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
+import java.util.Set;
 import java.util.stream.Collectors;
-import java.util.List;
+
 @Slf4j
 @Service
-public class UserService {
-
-
-    private final UserRepository userRepository;
+public class UserService implements UserServiceInterface {
 
     @Autowired
-    public UserService(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
+    private UserRepository userRepository;
 
     @Transactional
+    @Override
     public UserResponseDTO saveUser(User user) {
         log.info("Saving user: {}", user);
         return mapToUserResponseDTO(userRepository.save(user));
     }
 
+    @Override
     public Optional<UserResponseDTO> getUserById(Long userId) {
+        log.info("Getting user by ID: {}", userId);
+        if (!userRepository.existsById(userId)) {
+            throw new ResourceNotFoundException("User not found with ID: " + userId);
+        }
         return userRepository.findById(userId).map(this::mapToUserResponseDTO);
     }
 
-    public UserResponseDTO getUserByEmail(String email) {
-        User user = userRepository.findByEmail(email);
-        if (user == null) {
-            throw new IllegalArgumentException("User with email " + email + " not found.");
-        }
-        return mapToUserResponseDTO(user);
-    }
 
+    @Override
     @Transactional
     public void deleteUser(Long userId) {
+        log.info("Deleting user by ID: {}", userId);
+        if(!userRepository.existsById(userId)) {
+            throw new ResourceNotFoundException("User not found with ID: " + userId);
+        }
         userRepository.deleteById(userId);
     }
 
-    public List<UserResponseDTO> getUsers() {
+    @Override
+    public Set<UserResponseDTO> getUsers() {
+        log.info("Getting all users");
+        if (userRepository.count() == 0) {
+            throw new ResourceNotFoundException("No users found");
+        }
         return userRepository.findAll()
                 .stream()
                 .map(this::mapToUserResponseDTO)
-                .collect(Collectors.toList());
+                .collect(Collectors.toSet());
     }
 
     private UserResponseDTO mapToUserResponseDTO(User user) {
